backend/requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
requests==2.28.1
amadeus==5.0.0
pytest==7.4.0
pytest-cov==4.1.0
